+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add menu action slots to manage plugins page
+- Add a share menu action for exporting Asset Link plugins
 
 ## [1.0.0-alpha5] - 2023-03-14